add switchScrollingEffect file
